+++ Lib/PvpXray/Verifier.cs
@@ -6,21 +6,21 @@ using System.Text;
 
 namespace PvpXray
 {
-    public class Validator
+    public class Verifier
     {
-        /// All PVP checks implemented by Validator.
+        /// All PVP checks implemented by Verifier.
         public static readonly IReadOnlyList<string> Checks =
             Context.Checks
-                .Concat(ChangelogValidations.Checks)
-                .Concat(DocumentationValidations.Checks)
-                .Concat(LicenseValidations.Checks)
-                .Concat(ManifestTypeValidations.Checks)
-                .Concat(ManifestValidations.Checks)
-                .Concat(MetaFileValidations.Checks)
-                .Concat(MetaGuidValidations.Checks)
-                .Concat(PathValidations.Checks)
-                .Concat(SampleValidations.Checks)
-                .Concat(ThirdPartyNoticesValidations.Checks)
+                .Concat(ChangelogVerifier.Checks)
+                .Concat(DocumentationVerifier.Checks)
+                .Concat(LicenseVerifier.Checks)
+                .Concat(ManifestTypeVerifier.Checks)
+                .Concat(ManifestVerifier.Checks)
+                .Concat(MetaFileVerifier.Checks)
+                .Concat(MetaGuidVerifier.Checks)
+                .Concat(PathVerifier.Checks)
+                .Concat(SampleVerifier.Checks)
+                .Concat(ThirdPartyNoticesVerifier.Checks)
                 .OrderBy(s => s, StringComparer.Ordinal)
                 .ToArray();
 
@@ -40,16 +40,22 @@ namespace PvpXray
             public IReadOnlyList<string> Files { get; }
 
             readonly Action<string, string> m_AddError;
+            readonly Action<string, string> m_Skip;
             readonly HashSet<string> m_CurrentBatchChecks;
             readonly Json m_Manifest;
             readonly string m_ManifestError;
             readonly IPackage m_Package;
             readonly HashSet<string> m_PreviousErrors;
+            // TODO: Remove preprocessor directive when we have at least one check using the Skip method.
+#pragma warning disable CS0649
+            readonly Dictionary<string, string> m_PreviousSkipReasons;
+#pragma warning restore CS0649
 
-            public Context(IPackage package, Action<string, string> addError)
+            public Context(IPackage package, Action<string, string> addError, Action<string, string> skip)
             {
                 m_Package = package;
                 m_AddError = addError;
+                m_Skip = skip;
                 m_CurrentBatchChecks = new HashSet<string>();
                 m_PreviousErrors = new HashSet<string>();
 
@@ -59,7 +65,7 @@ namespace PvpXray
                 {
                     var text = ReadFileToString("package.json");
 
-                    // UTF-8 BOM is unwelcome, but we can proceed with validation.
+                    // UTF-8 BOM is unwelcome, but we can proceed with verification.
                     if (text.StartsWithOrdinal("\ufeff"))
                     {
                         addError("PVP-100-1", "manifest file contains UTF-8 BOM");
@@ -89,6 +95,21 @@ namespace PvpXray
 
                     m_AddError(checkId, error);
                 }
+            }
+
+            public void Skip(string checkId, string reason)
+            {
+                if (m_PreviousSkipReasons.TryGetValue(checkId, out var previousReason))
+                {
+                    throw new InvalidOperationException($"batch tried to skip check ID {checkId} more than once; previous skip reason: {previousReason}, new skip reason: {reason}");
+                }
+
+                if (!m_CurrentBatchChecks.Contains(checkId))
+                {
+                    throw new InvalidOperationException($"batch tried to skip undeclared check ID {checkId}: {reason}");
+                }
+
+                m_Skip(checkId, reason);
             }
 
             /// Determine if the specified path exists in package and is a directory. Note that
@@ -140,7 +161,7 @@ namespace PvpXray
 
             public void RunBatch(string[] checks, Action<Context> runner)
             {
-                var undeclaredChecks = checks.Except(Validator.Checks).ToArray();
+                var undeclaredChecks = checks.Except(Verifier.Checks).ToArray();
                 if (undeclaredChecks.Any())
                 {
                     throw new InvalidOperationException($"batch run for undeclared checks: {string.Join(", ", undeclaredChecks)}");
@@ -166,19 +187,22 @@ namespace PvpXray
             }
         }
 
-        public void Validate(IPackage package, Action<string, string> addError)
+        public void Verify(IPackage package, Action<string, string> addError, Action<string, string> skip, IPvpHttpClient httpClient)
         {
-            var context = new Context(package, addError);
-            context.RunBatch(ChangelogValidations.Checks, ChangelogValidations.Run);
-            context.RunBatch(LicenseValidations.Checks, LicenseValidations.Run);
-            context.RunBatch(ManifestTypeValidations.Checks, ManifestTypeValidations.Run);
-            context.RunBatch(ManifestValidations.Checks, ManifestValidations.Run);
-            context.RunBatch(MetaFileValidations.Checks, MetaFileValidations.Run);
-            context.RunBatch(MetaGuidValidations.Checks, MetaGuidValidations.Run);
-            context.RunBatch(PathValidations.Checks, PathValidations.Run);
-            context.RunBatch(SampleValidations.Checks, SampleValidations.Run);
-            context.RunBatch(ThirdPartyNoticesValidations.Checks, ThirdPartyNoticesValidations.Run);
-            context.RunBatch(DocumentationValidations.Checks, DocumentationValidations.Run);
+            // TODO: New checks that require access to a network resource must be skipped immediately with reason "offline_requested" if httpClient is null.
+            // TODO: New checks that call httpClient.GetStream must be skipped with reason "network_error" if it throws PvpHttpException.
+
+            var context = new Context(package, addError, skip);
+            context.RunBatch(ChangelogVerifier.Checks, ChangelogVerifier.Run);
+            context.RunBatch(LicenseVerifier.Checks, LicenseVerifier.Run);
+            context.RunBatch(ManifestTypeVerifier.Checks, ManifestTypeVerifier.Run);
+            context.RunBatch(ManifestVerifier.Checks, ManifestVerifier.Run);
+            context.RunBatch(MetaFileVerifier.Checks, MetaFileVerifier.Run);
+            context.RunBatch(MetaGuidVerifier.Checks, MetaGuidVerifier.Run);
+            context.RunBatch(PathVerifier.Checks, PathVerifier.Run);
+            context.RunBatch(SampleVerifier.Checks, SampleVerifier.Run);
+            context.RunBatch(ThirdPartyNoticesVerifier.Checks, ThirdPartyNoticesVerifier.Run);
+            context.RunBatch(DocumentationVerifier.Checks, DocumentationVerifier.Run);
         }
     }
 }